[Keychron C1 Series] Reorganizing keymaps
C1 Plain still broken BTW XD
PR ain't ready until C1 Plain gets fixed.
