+++ keyboards/keychron/c1/plain/keymaps/via/keymap.c
@@ -1,3 +1,3 @@
 // Include default keymap
 // This way the VIA keymap is always up to date with the default keymap
-#include "../default/keymap.c"
+#include "../../../keymaps/default/keymap.c"
