+++ keyboards/keychron/c1/readme.md
@@ -9,65 +9,6 @@ A customizable TKL keyboard.
 * Hardware Availability: [Keychron](https://www.keychron.com/products/keychron-c1-wired-mechanical-keyboard)
 
 * * *
-# Keymapping
-### Windows Mode
-<details>
-
-Without Fn | With Fn
----------- | -------
-F1 | Decrease PC Brightness
-F2 | Increase PC Brightness
-F3 | Task View
-F4 | File Explorer
-F5 | Decrease KB Brightness
-F6 | Increase KB Brightness
-F7 | Previous Track
-F8 | Play/Pause Track
-F9 | Next Track
-F10 | Mute
-F11 | Decrease Volume
-F12 | Increase Volume
-Print Screen | Snip & Sketch
-Cortana / Microsoft Teams | None
-Change RGB | Toggle RGB
-⊞ | Windows Key Lock Toggle
-Home | NKRO Toggle
-↑ | Increase RGB Hue
-← | Decrease RGB Saturation
-↓ | Decrease RGB Hue
-→ | Increase RGB Saturation
-
-</details>
-
-### Mac Mode
-<details>
-
-Without Fn | With Fn
----------- | -------
-F1 | Decrease PC Brightness
-F2 | Increase PC Brightness
-F3 | Mission Control
-F4 | Finder
-F5 | Decrease KB Brightness
-F6 | Increase KB Brightness
-F7 | Previous Track
-F8 | Play/Pause Track
-F9 | Next Track
-F10 | Mute
-F11 | Decrease Volume
-F12 | Increase Volume
-Take a screenshot (whole screen) | Take a screenshot (specific area)
-Siri (hold down) | None
-Change RGB | Toggle RGB
-Home | NKRO Toggle
-↑ | Increase RGB Hue
-← | Decrease RGB Saturation
-↓ | Decrease RGB Hue
-→ | Increase RGB Saturation
-
-</details>
-
-* * *
 
 For more information and more detailed flashing instructions, please visit https://github.com/CanUnesi/QMK-on-K6#readme
